Port integration tests to archery
Work on dockerization, refine  CLI

Use archery in verify-release-candidate.sh

Less aggressive parallelization

Add missing license

Fix hadolint warning. Force integration to build to test Travis

Try explicit pull

More tweaks

Build/pull cpp also

Do not use docker-compose build

Restore detect-changes logic for integration test entry

Address review comments
