+++ .travis.yml
@@ -185,33 +185,15 @@ matrix:
     - docker-compose run java-all-jdks
 
   # -------------------------------------------------------------------------
-  - name: "Integration w/ OpenJDK 8, conda-forge toolchain"
-    language: java
+  - name: "Protocol / Flight Integration Tests"
+    language: cpp
     os: linux
     env: ARROW_TEST_GROUP=integration
-    jdk: openjdk8
-    env:
-    - ARROW_TRAVIS_USE_TOOLCHAIN=1
-    - ARROW_TRAVIS_FLIGHT=1
-    - ARROW_TRAVIS_INTEGRATION=1
-    # TODO(ARROW-4764): Using system java and plasma built with conda doesn't work
-    #- ARROW_TRAVIS_PLASMA=1
-    #- ARROW_TRAVIS_PLASMA_JAVA_CLIENT=1
-    - ARROW_TRAVIS_MIMALLOC=1
     before_script:
     - if [ $ARROW_CI_INTEGRATION_AFFECTED != "1" ]; then exit; fi
-    - $TRAVIS_BUILD_DIR/ci/travis_install_clang_tools.sh
-    - $TRAVIS_BUILD_DIR/ci/travis_install_linux.sh
-    # C++: Only build integration targets
-    - export ARROW_CPP_BUILD_TARGETS=arrow-integration
-    - export ARROW_TRAVIS_OPTIONAL_INSTALL=1
-    - $TRAVIS_BUILD_DIR/ci/travis_before_script_cpp.sh
-    # Javascript
-    - nvm install 11.12
-    - $TRAVIS_BUILD_DIR/ci/travis_before_script_js.sh
+    - docker-compose pull integration
     script:
-    - $TRAVIS_BUILD_DIR/ci/travis_script_integration.sh
-    # - $TRAVIS_BUILD_DIR/ci/travis_script_plasma_java_client.sh
+    - docker-compose run integration
 
   # -------------------------------------------------------------------------
   - name: "NodeJS"
