+++ dev/archery/archery/integration/datagen.py
@@ -16,92 +16,15 @@
 # under the License.
 
 from collections import OrderedDict
-import argparse
 import binascii
-import contextlib
-import glob
-import gzip
-import itertools
 import json
 import os
 import random
-import six
-import string
-import subprocess
-import sys
 import tempfile
-import traceback
-import uuid
-import errno
 
 import numpy as np
 
-
-ARROW_HOME = os.path.abspath(__file__).rsplit("/", 2)[0]
-
-# Control for flakiness
-np.random.seed(12345)
-
-
-def load_version_from_pom():
-    import xml.etree.ElementTree as ET
-    tree = ET.parse(os.path.join(ARROW_HOME, 'java', 'pom.xml'))
-    tag_pattern = '{http://maven.apache.org/POM/4.0.0}version'
-    version_tag = list(tree.getroot().findall(tag_pattern))[0]
-    return version_tag.text
-
-
-def guid():
-    return uuid.uuid4().hex
-
-
-# from pandas
-RANDS_CHARS = np.array(list(string.ascii_letters + string.digits),
-                       dtype=(np.str_, 1))
-
-
-def rands(nchars):
-    """
-    Generate one random byte string.
-
-    See `rands_array` if you want to create an array of random strings.
-
-    """
-    return ''.join(np.random.choice(RANDS_CHARS, nchars))
-
-
-def tobytes(o):
-    if isinstance(o, six.text_type):
-        return o.encode('utf8')
-    return o
-
-
-def frombytes(o):
-    if isinstance(o, six.binary_type):
-        return o.decode('utf8')
-    return o
-
-
-# from the merge_arrow_pr.py script
-def run_cmd(cmd):
-    if isinstance(cmd, six.string_types):
-        cmd = cmd.split(' ')
-
-    try:
-        output = subprocess.check_output(cmd, stderr=subprocess.STDOUT)
-    except subprocess.CalledProcessError as e:
-        # this avoids hiding the stdout / stderr of failed processes
-        print('Command failed: %s' % ' '.join(cmd))
-        print('With output:')
-        print('--------------')
-        print(frombytes(e.output))
-        print('--------------')
-        raise e
-
-    return frombytes(output)
-
-# ----------------------------------------------------------------------
-# Data generation
+from .util import (frombytes, rands, tobytes, SKIP_ARROW, SKIP_FLIGHT)
 
 
 class DataType(object):
@@ -910,11 +833,6 @@ class JsonRecordBatch(object):
         ])
 
 
-# SKIP categories
-SKIP_ARROW = 'arrow'
-SKIP_FLIGHT = 'flight'
-
-
 class JsonFile(object):
 
     def __init__(self, name, schema, batches, dictionaries=None,
@@ -1179,616 +1097,3 @@ def get_generated_json_files(tempdir=None, flight=False):
         generated_paths.append(file_obj)
 
     return generated_paths
-
-
-# ----------------------------------------------------------------------
-# Testing harness
-
-
-class IntegrationRunner(object):
-
-    def __init__(self, json_files, testers, args):
-        self.json_files = json_files
-        self.testers = testers
-        self.temp_dir = args.tempdir or tempfile.mkdtemp()
-        self.debug = args.debug
-        self.stop_on_error = args.stop_on_error
-        self.gold_dirs = args.gold_dirs
-
-    def run(self):
-        failures = []
-
-        for producer, consumer in itertools.product(
-                filter(lambda t: t.PRODUCER, self.testers),
-                filter(lambda t: t.CONSUMER, self.testers)):
-            for failure in self._compare_implementations(
-              producer, consumer, self._produce_consume, self.json_files):
-                failures.append(failure)
-        if self.gold_dirs:
-            for gold_dir, consumer in itertools.product(
-                    self.gold_dirs,
-                    filter(lambda t: t.CONSUMER, self.testers)):
-                print('\n\n\n\n')
-                print('******************************************************')
-                print('Tests against golden files in {}'.format(gold_dir))
-                print('******************************************************')
-
-                def run_gold(producer, consumer, test_case):
-                    self._run_gold(gold_dir, producer, consumer, test_case)
-                for failure in self._compare_implementations(
-                  consumer, consumer,  run_gold, self._gold_tests(gold_dir)):
-                    failures.append(failure)
-
-        return failures
-
-    def run_flight(self):
-        failures = []
-        servers = filter(lambda t: t.FLIGHT_SERVER, self.testers)
-        clients = filter(lambda t: (t.FLIGHT_CLIENT and t.CONSUMER),
-                         self.testers)
-        for server, client in itertools.product(servers, clients):
-            for failure in self._compare_flight_implementations(server,
-                                                                client):
-                failures.append(failure)
-        return failures
-
-    def _gold_tests(self, gold_dir):
-        prefix = os.path.basename(os.path.normpath(gold_dir))
-        SUFFIX = ".json.gz"
-        golds = [jf for jf in os.listdir(gold_dir) if jf.endswith(SUFFIX)]
-        for json_path in golds:
-            name = json_path[json_path.index('_')+1: -len(SUFFIX)]
-            base_name = prefix + "_" + name + ".gold.json"
-            out_path = os.path.join(self.temp_dir, base_name)
-            with gzip.open(os.path.join(gold_dir, json_path)) as i:
-                with open(out_path, "wb") as out:
-                    out.write(i.read())
-
-            try:
-                skip = next(f for f in self.json_files
-                            if f.name == name).skip
-            except StopIteration:
-                skip = set()
-            yield JsonFile(name, None, None, skip=skip, path=out_path)
-
-    def _compare_implementations(
-            self, producer, consumer, run_binaries, test_cases):
-        print('##########################################################')
-        print(
-            '{0} producing, {1} consuming'.format(producer.name, consumer.name)
-        )
-        print('##########################################################')
-
-        for test_case in test_cases:
-            json_path = test_case.path
-            print('==========================================================')
-            print('Testing file {0}'.format(json_path))
-            print('==========================================================')
-
-            if producer.name in test_case.skip:
-                print('-- Skipping test because producer {0} does '
-                      'not support'.format(producer.name))
-                continue
-
-            if consumer.name in test_case.skip:
-                print('-- Skipping test because consumer {0} does '
-                      'not support'.format(consumer.name))
-                continue
-
-            if SKIP_ARROW in test_case.skip:
-                print('-- Skipping test')
-                continue
-
-            try:
-                run_binaries(producer, consumer, test_case)
-            except Exception:
-                traceback.print_exc()
-                yield (test_case, producer, consumer, sys.exc_info())
-                if self.stop_on_error:
-                    break
-                else:
-                    continue
-
-    def _produce_consume(self, producer, consumer, test_case):
-        # Make the random access file
-        json_path = test_case.path
-        file_id = guid()[:8]
-        name = os.path.splitext(os.path.basename(json_path))[0]
-
-        producer_file_path = os.path.join(self.temp_dir, file_id + '_' +
-                                          name + '.json_as_file')
-        producer_stream_path = os.path.join(self.temp_dir, file_id + '_' +
-                                            name + '.producer_file_as_stream')
-        consumer_file_path = os.path.join(self.temp_dir, file_id + '_' +
-                                          name + '.consumer_stream_as_file')
-
-        print('-- Creating binary inputs')
-        producer.json_to_file(json_path, producer_file_path)
-
-        # Validate the file
-        print('-- Validating file')
-        consumer.validate(json_path, producer_file_path)
-
-        print('-- Validating stream')
-        producer.file_to_stream(producer_file_path, producer_stream_path)
-        consumer.stream_to_file(producer_stream_path, consumer_file_path)
-        consumer.validate(json_path, consumer_file_path)
-
-    def _run_gold(self, gold_dir, producer, consumer, test_case):
-        json_path = test_case.path
-
-        # Validate the file
-        print('-- Validating file')
-        producer_file_path = os.path.join(
-            gold_dir, "generated_" + test_case.name + ".arrow_file")
-        consumer.validate(json_path, producer_file_path)
-
-        print('-- Validating stream')
-        consumer_stream_path = os.path.join(
-            gold_dir, "generated_" + test_case.name + ".stream")
-        file_id = guid()[:8]
-        name = os.path.splitext(os.path.basename(json_path))[0]
-
-        consumer_file_path = os.path.join(self.temp_dir, file_id + '_' +
-                                          name + '.consumer_stream_as_file')
-
-        consumer.stream_to_file(consumer_stream_path, consumer_file_path)
-        consumer.validate(json_path, consumer_file_path)
-
-    def _compare_flight_implementations(self, producer, consumer):
-        print('##########################################################')
-        print(
-            '{0} serving, {1} requesting'.format(producer.name, consumer.name)
-        )
-        print('##########################################################')
-
-        for test_case in self.json_files:
-            json_path = test_case.path
-            print('=' * 58)
-            print('Testing file {0}'.format(json_path))
-            print('=' * 58)
-
-            if ('Java' in (producer.name, consumer.name) and
-               "map" in test_case.name):
-                print('TODO(ARROW-1279): Enable map tests ' +
-                      ' for Java and JS once Java supports them and JS\'' +
-                      ' are unbroken')
-                continue
-
-            if SKIP_FLIGHT in test_case.skip:
-                print('-- Skipping test')
-                continue
-
-            try:
-                with producer.flight_server():
-                    # Have the client upload the file, then download and
-                    # compare
-                    consumer.flight_request(producer.FLIGHT_PORT, json_path)
-            except Exception:
-                traceback.print_exc()
-                yield (test_case, producer, consumer, sys.exc_info())
-                continue
-
-
-class Tester(object):
-    PRODUCER = False
-    CONSUMER = False
-    FLIGHT_SERVER = False
-    FLIGHT_CLIENT = False
-    FLIGHT_PORT = 31337
-
-    def __init__(self, args):
-        self.args = args
-        self.debug = args.debug
-
-    def run_shell_command(self, cmd):
-        cmd = ' '.join(cmd)
-        if self.debug:
-            print(cmd)
-        subprocess.check_call(cmd, shell=True)
-
-    def json_to_file(self, json_path, arrow_path):
-        raise NotImplementedError
-
-    def stream_to_file(self, stream_path, file_path):
-        raise NotImplementedError
-
-    def file_to_stream(self, file_path, stream_path):
-        raise NotImplementedError
-
-    def validate(self, json_path, arrow_path):
-        raise NotImplementedError
-
-    def flight_server(self):
-        raise NotImplementedError
-
-    def flight_request(self, port, json_path):
-        raise NotImplementedError
-
-
-class JavaTester(Tester):
-    PRODUCER = True
-    CONSUMER = True
-    FLIGHT_SERVER = True
-    FLIGHT_CLIENT = True
-
-    FLIGHT_PORT = 31338
-
-    JAVA_OPTS = ['-Dio.netty.tryReflectionSetAccessible=true']
-
-    _arrow_version = load_version_from_pom()
-    ARROW_TOOLS_JAR = os.environ.get(
-        'ARROW_JAVA_INTEGRATION_JAR',
-        os.path.join(ARROW_HOME,
-                     'java/tools/target/arrow-tools-{}-'
-                     'jar-with-dependencies.jar'.format(_arrow_version)))
-    ARROW_FLIGHT_JAR = os.environ.get(
-        'ARROW_FLIGHT_JAVA_INTEGRATION_JAR',
-        os.path.join(ARROW_HOME,
-                     'java/flight/target/arrow-flight-{}-'
-                     'jar-with-dependencies.jar'.format(_arrow_version)))
-    ARROW_FLIGHT_SERVER = ('org.apache.arrow.flight.example.integration.'
-                           'IntegrationTestServer')
-    ARROW_FLIGHT_CLIENT = ('org.apache.arrow.flight.example.integration.'
-                           'IntegrationTestClient')
-
-    name = 'Java'
-
-    def _run(self, arrow_path=None, json_path=None, command='VALIDATE'):
-        cmd = ['java'] + self.JAVA_OPTS + \
-            ['-cp', self.ARROW_TOOLS_JAR, 'org.apache.arrow.tools.Integration']
-
-        if arrow_path is not None:
-            cmd.extend(['-a', arrow_path])
-
-        if json_path is not None:
-            cmd.extend(['-j', json_path])
-
-        cmd.extend(['-c', command])
-
-        if self.debug:
-            print(' '.join(cmd))
-
-        run_cmd(cmd)
-
-    def validate(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'VALIDATE')
-
-    def json_to_file(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'JSON_TO_ARROW')
-
-    def stream_to_file(self, stream_path, file_path):
-        cmd = ['java'] + self.JAVA_OPTS + \
-            ['-cp', self.ARROW_TOOLS_JAR,
-             'org.apache.arrow.tools.StreamToFile', stream_path, file_path]
-        if self.debug:
-            print(' '.join(cmd))
-        run_cmd(cmd)
-
-    def file_to_stream(self, file_path, stream_path):
-        cmd = ['java'] + self.JAVA_OPTS + \
-            ['-cp', self.ARROW_TOOLS_JAR,
-             'org.apache.arrow.tools.FileToStream', file_path, stream_path]
-        if self.debug:
-            print(' '.join(cmd))
-        run_cmd(cmd)
-
-    def flight_request(self, port, json_path):
-        cmd = ['java'] + self.JAVA_OPTS + \
-            ['-cp', self.ARROW_FLIGHT_JAR, self.ARROW_FLIGHT_CLIENT,
-             '-port', str(port), '-j', json_path]
-        if self.debug:
-            print(' '.join(cmd))
-        run_cmd(cmd)
-
-    @contextlib.contextmanager
-    def flight_server(self):
-        cmd = ['java'] + self.JAVA_OPTS + \
-            ['-cp', self.ARROW_FLIGHT_JAR, self.ARROW_FLIGHT_SERVER,
-             '-port', str(self.FLIGHT_PORT)]
-        if self.debug:
-            print(' '.join(cmd))
-        server = subprocess.Popen(cmd, stdout=subprocess.PIPE)
-        try:
-            output = server.stdout.readline().decode()
-            if not output.startswith("Server listening on localhost"):
-                raise RuntimeError(
-                    "Flight-Java server did not start properly, output: " +
-                    output)
-            yield
-        finally:
-            server.kill()
-            server.wait(5)
-
-
-class CPPTester(Tester):
-    PRODUCER = True
-    CONSUMER = True
-    FLIGHT_SERVER = True
-    FLIGHT_CLIENT = True
-
-    EXE_PATH = os.environ.get(
-        'ARROW_CPP_EXE_PATH',
-        os.path.join(ARROW_HOME, 'cpp/build/debug'))
-
-    CPP_INTEGRATION_EXE = os.path.join(EXE_PATH, 'arrow-json-integration-test')
-    STREAM_TO_FILE = os.path.join(EXE_PATH, 'arrow-stream-to-file')
-    FILE_TO_STREAM = os.path.join(EXE_PATH, 'arrow-file-to-stream')
-
-    FLIGHT_PORT = 31337
-
-    FLIGHT_SERVER_CMD = [
-        os.path.join(EXE_PATH, 'flight-test-integration-server'),
-        "-port", str(FLIGHT_PORT)]
-    FLIGHT_CLIENT_CMD = [
-        os.path.join(EXE_PATH, 'flight-test-integration-client'),
-        "-host", "localhost"]
-
-    name = 'C++'
-
-    def _run(self, arrow_path=None, json_path=None, command='VALIDATE'):
-        cmd = [self.CPP_INTEGRATION_EXE, '--integration']
-
-        if arrow_path is not None:
-            cmd.append('--arrow=' + arrow_path)
-
-        if json_path is not None:
-            cmd.append('--json=' + json_path)
-
-        cmd.append('--mode=' + command)
-
-        if self.debug:
-            print(' '.join(cmd))
-
-        run_cmd(cmd)
-
-    def validate(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'VALIDATE')
-
-    def json_to_file(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'JSON_TO_ARROW')
-
-    def stream_to_file(self, stream_path, file_path):
-        cmd = ['cat', stream_path, '|', self.STREAM_TO_FILE, '>', file_path]
-        self.run_shell_command(cmd)
-
-    def file_to_stream(self, file_path, stream_path):
-        cmd = [self.FILE_TO_STREAM, file_path, '>', stream_path]
-        self.run_shell_command(cmd)
-
-    @contextlib.contextmanager
-    def flight_server(self):
-        if self.debug:
-            print(' '.join(self.FLIGHT_SERVER_CMD))
-        server = subprocess.Popen(self.FLIGHT_SERVER_CMD,
-                                  stdout=subprocess.PIPE)
-        try:
-            output = server.stdout.readline().decode()
-            if not output.startswith("Server listening on localhost"):
-                raise RuntimeError(
-                    "Flight-C++ server did not start properly, output: " +
-                    output)
-            yield
-        finally:
-            server.kill()
-            server.wait(5)
-
-    def flight_request(self, port, json_path):
-        cmd = self.FLIGHT_CLIENT_CMD + [
-            '-port=' + str(port),
-            '-path=' + json_path,
-        ]
-        if self.debug:
-            print(' '.join(cmd))
-        run_cmd(cmd)
-
-
-class JSTester(Tester):
-    PRODUCER = True
-    CONSUMER = True
-
-    EXE_PATH = os.path.join(ARROW_HOME, 'js/bin')
-    VALIDATE = os.path.join(EXE_PATH, 'integration.js')
-    JSON_TO_ARROW = os.path.join(EXE_PATH, 'json-to-arrow.js')
-    STREAM_TO_FILE = os.path.join(EXE_PATH, 'stream-to-file.js')
-    FILE_TO_STREAM = os.path.join(EXE_PATH, 'file-to-stream.js')
-
-    name = 'JS'
-
-    def _run(self, exe_cmd, arrow_path=None, json_path=None,
-             command='VALIDATE'):
-        cmd = [exe_cmd]
-
-        if arrow_path is not None:
-            cmd.extend(['-a', arrow_path])
-
-        if json_path is not None:
-            cmd.extend(['-j', json_path])
-
-        cmd.extend(['--mode', command])
-
-        if self.debug:
-            print(' '.join(cmd))
-
-        run_cmd(cmd)
-
-    def validate(self, json_path, arrow_path):
-        return self._run(self.VALIDATE, arrow_path, json_path, 'VALIDATE')
-
-    def json_to_file(self, json_path, arrow_path):
-        cmd = ['node',
-               '--no-warnings', self.JSON_TO_ARROW,
-               '-a', arrow_path,
-               '-j', json_path]
-        self.run_shell_command(cmd)
-
-    def stream_to_file(self, stream_path, file_path):
-        cmd = ['cat', stream_path, '|',
-               'node', '--no-warnings', self.STREAM_TO_FILE, '>',
-               file_path]
-        self.run_shell_command(cmd)
-
-    def file_to_stream(self, file_path, stream_path):
-        cmd = ['cat', file_path, '|',
-               'node', '--no-warnings', self.FILE_TO_STREAM, '>',
-               stream_path]
-        self.run_shell_command(cmd)
-
-
-class GoTester(Tester):
-    PRODUCER = True
-    CONSUMER = True
-
-    # FIXME(sbinet): revisit for Go modules
-    GOPATH = os.getenv('GOPATH', '~/go')
-    GOBIN = os.environ.get('GOBIN', os.path.join(GOPATH, 'bin'))
-
-    GO_INTEGRATION_EXE = os.path.join(GOBIN, 'arrow-json-integration-test')
-    STREAM_TO_FILE = os.path.join(GOBIN, 'arrow-stream-to-file')
-    FILE_TO_STREAM = os.path.join(GOBIN, 'arrow-file-to-stream')
-
-    name = 'Go'
-
-    def _run(self, arrow_path=None, json_path=None, command='VALIDATE'):
-        cmd = [self.GO_INTEGRATION_EXE]
-
-        if arrow_path is not None:
-            cmd.extend(['-arrow', arrow_path])
-
-        if json_path is not None:
-            cmd.extend(['-json', json_path])
-
-        cmd.extend(['-mode', command])
-
-        if self.debug:
-            print(' '.join(cmd))
-
-        run_cmd(cmd)
-
-    def validate(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'VALIDATE')
-
-    def json_to_file(self, json_path, arrow_path):
-        return self._run(arrow_path, json_path, 'JSON_TO_ARROW')
-
-    def stream_to_file(self, stream_path, file_path):
-        cmd = ['cat', stream_path, '|', self.STREAM_TO_FILE, '>', file_path]
-        self.run_shell_command(cmd)
-
-    def file_to_stream(self, file_path, stream_path):
-        cmd = [self.FILE_TO_STREAM, file_path, '>', stream_path]
-        self.run_shell_command(cmd)
-
-
-def get_static_json_files():
-    glob_pattern = os.path.join(ARROW_HOME, 'integration', 'data', '*.json')
-    return [JsonFile(name=os.path.basename(p), path=p, skip=set(),
-                     schema=None, batches=None)
-            for p in glob.glob(glob_pattern)]
-
-
-def run_all_tests(args):
-    testers = []
-
-    if args.enable_cpp:
-        testers.append(CPPTester(args))
-
-    if args.enable_java:
-        testers.append(JavaTester(args))
-
-    if args.enable_js:
-        testers.append(JSTester(args))
-
-    if args.enable_go:
-        testers.append(GoTester(args))
-
-    static_json_files = get_static_json_files()
-    generated_json_files = get_generated_json_files(tempdir=args.tempdir,
-                                                    flight=args.run_flight)
-    json_files = static_json_files + generated_json_files
-
-    runner = IntegrationRunner(json_files, testers, args)
-    failures = []
-    failures.extend(runner.run())
-    if args.run_flight:
-        failures.extend(runner.run_flight())
-
-    fail_count = 0
-    if failures:
-        print("################# FAILURES #################")
-        for test_case, producer, consumer, exc_info in failures:
-            fail_count += 1
-            print("FAILED TEST:", end=" ")
-            print(test_case.name, producer.name, "producing, ",
-                  consumer.name, "consuming")
-            if exc_info:
-                traceback.print_exception(*exc_info)
-            print()
-
-    print(fail_count, "failures")
-    if fail_count > 0:
-        sys.exit(1)
-
-
-def write_js_test_json(directory):
-    generate_map_case().write(os.path.join(directory, 'map.json'))
-    generate_nested_case().write(os.path.join(directory, 'nested.json'))
-    generate_decimal_case().write(os.path.join(directory, 'decimal.json'))
-    generate_datetime_case().write(os.path.join(directory, 'datetime.json'))
-    (generate_dictionary_case()
-     .write(os.path.join(directory, 'dictionary.json')))
-    (generate_primitive_case([])
-     .write(os.path.join(directory, 'primitive_no_batches.json')))
-    (generate_primitive_case([7, 10])
-     .write(os.path.join(directory, 'primitive.json')))
-    (generate_primitive_case([0, 0, 0])
-     .write(os.path.join(directory, 'primitive-empty.json')))
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Arrow integration test CLI')
-
-    parser.add_argument('--enable-c++', dest='enable_cpp',
-                        action='store', type=int, default=1,
-                        help='Include C++ in integration tests')
-    parser.add_argument('--enable-java', dest='enable_java',
-                        action='store', type=int, default=1,
-                        help='Include Java in integration tests')
-    parser.add_argument('--enable-js', dest='enable_js',
-                        action='store', type=int, default=1,
-                        help='Include JavaScript in integration tests')
-    parser.add_argument('--enable-go', dest='enable_go',
-                        action='store', type=int, default=1,
-                        help='Include Go in integration tests')
-
-    parser.add_argument('--write_generated_json', dest='generated_json_path',
-                        action='store', default=False,
-                        help='Generate test JSON')
-    parser.add_argument('--run_flight', dest='run_flight',
-                        action='store_true', default=False,
-                        help='Run Flight integration tests')
-    parser.add_argument('--debug', dest='debug', action='store_true',
-                        default=False,
-                        help='Run executables in debug mode as relevant')
-    parser.add_argument('--tempdir', dest='tempdir',
-                        default=tempfile.mkdtemp(),
-                        help=('Directory to use for writing '
-                              'integration test temporary files'))
-
-    parser.add_argument('-x', '--stop-on-error', dest='stop_on_error',
-                        action='store_true', default=False,
-                        help='Stop on first error')
-
-    parser.add_argument('--gold_dirs', action='append',
-                        help="gold integration test file paths")
-
-    args = parser.parse_args()
-    if args.generated_json_path:
-        try:
-            os.makedirs(args.generated_json_path)
-        except OSError as e:
-            if e.errno != errno.EEXIST:
-                raise
-        write_js_test_json(args.generated_json_path)
-    else:
-        run_all_tests(args)
